# src/app/api/openapi.yaml
openapi: 3.0.3
info:
  title: Stitch AI Manager API
  description: |
    REST API for programmatic control of Stitch workflows through natural language.
    
    The AI Manager API enables developers, CLI tools, and AI agents to create, modify, 
    and execute workflows without manual canvas editing.
    
    ## Features
    - Canvas CRUD operations (JSON and Mermaid formats)
    - Workflow execution and status monitoring
    - AI-powered natural language interface
    - Automatic version snapshots
    - Entity movement tracking
    
    ## Authentication
    All endpoints require authentication (implementation-specific).
    
  version: 1.0.0
  contact:
    name: Stitch API Support
  license:
    name: MIT
servers:
  - url: https://your-domain.com/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
...
paths:
...
  /canvas/{id}/status:
    get:
      tags:
        - Workflow
      summary: Get workflow status
      description: |
        Query the status of a workflow execution.
        
        Returns:
        - Overall workflow status
        - Individual node states
        - Node outputs (for completed nodes)
        - Error details (for failed nodes)
        - Final outputs (for completed workflows)
      operationId: getWorkflowStatus
      parameters:
        - name: id
          in: path
          required: true
          description: Canvas UUID
          schema:
            type: string
            format: uuid
        - name: runId
          in: query
          required: true
          description: Run UUID from run response
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetStatusResponse'
              examples:
                running:
                  summary: Workflow running
                  value:
                    runId: "660e8400-e29b-41d4-a716-446655440000"
                    status: "running"
                    nodes:
                      node-1:
                        status: "completed"
                        output: { prompt: "Generate video" }
                      node-2:
                        status: "running"
                    statusUrl: "https://your-domain.com/api/canvas/550e8400/status?runId=660e8400"
                completed:
                  summary: Workflow completed
                  value:
                    runId: "660e8400-e29b-41d4-a716-446655440000"
                    status: "completed"
                    nodes:
                      node-1:
                        status: "completed"
                        output: { prompt: "Generate video" }
                      node-2:
                        status: "completed"
                        output: { video_url: "https://..." }
                    finalOutputs:
                      node-2:
                        video_url: "https://..."
                    statusUrl: "https://your-domain.com/api/canvas/550e8400/status?runId=660e8400"
                failed:
                  summary: Workflow failed
                  value:
                    runId: "660e8400-e29b-41d4-a716-446655440000"
                    status: "failed"
                    nodes:
                      node-1:
                        status: "completed"
                        output: { prompt: "Generate video" }
                      node-2:
                        status: "failed"
                        error: "API rate limit exceeded"
                    statusUrl: "https://your-domain.com/api/canvas/550e8400/status?runId=660e8400"
        '400':
          description: Bad request (missing runId or run doesn't belong to canvas)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Canvas or run not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
        - code
      properties:
        error:
          type: string
          description: Human-readable error message
        details:
          type: array
          items:
            type: string
          description: Additional error details
        code:
          type: string
          enum:
            - BAD_REQUEST
            - NOT_FOUND
            - INTERNAL_ERROR
            - VALIDATION_ERROR
            - PARSE_ERROR
            - LLM_ERROR
          description: Error code for programmatic handling

    # Canvas Metadata
...
    NodeStateResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - pending
            - running
            - completed
            - failed
            - waiting_for_user
          description: Node status
        output:
          type: object
          description: Node output (for completed nodes)
        error:
          type: string
          description: Error message (for failed nodes)

    # Get Status Response
    GetStatusResponse:
      type: object
      required:
        - runId
        - status
        - nodes
        - statusUrl
      properties:
        runId:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - pending
            - running
            - completed
            - failed
          description: Overall workflow status
        nodes:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/NodeStateResponse'
          description: Map of node IDs to their states
        finalOutputs:
          type: object
          description: Outputs from terminal nodes (for completed workflows)
        statusUrl:
          type: string
          format: uri
          description: URL to continue polling

    # AI Manager Request
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []